# libs/contracts/rest/openapi.snapshot.json
{
  "components": {
    "schemas": {
      "BindSourceRequestDto": {
        "properties": {},
        "type": "object"
      },
      "ChangeSourceBindingStatusRequestDto": {
        "properties": {},
        "type": "object"
      },
...
    }
  },
  "info": {
    "contact": {},
    "description": "Backend/API-first social monitoring MVP.",
    "title": "Social Monitor API",
    "version": "0.1.0"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/topics/{topicId}/source-bindings": {
      "get": {
        "operationId": "SourceBindingController_list",
        "parameters": [
          {
            "in": "path",
            "name": "topicId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "type": "number"
            }
          },
          {
            "description": "Optional Bearer API key. Requires read:topics. If supplied, x-workspace-role is not required.",
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Comma-separated workspace roles. Source binding reads allow owner, admin, member or viewer. Required when Authorization bearer API key is not supplied.",
            "in": "header",
            "name": "x-workspace-role",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-workspace-id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-tenant-id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "List source bindings for a topic.",
        "tags": [
          "source-bindings"
        ]
      },
      "post": {
        "operationId": "SourceBindingController_create",
        "parameters": [
          {
            "in": "path",
            "name": "topicId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "idempotency-key",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Optional Bearer API key. Requires write:source_bindings. If supplied, x-workspace-role is not required.",
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Comma-separated workspace roles. Source binding creation requires owner or admin. Required when Authorization bearer API key is not supplied.",
            "in": "header",
            "name": "x-workspace-role",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-workspace-id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-tenant-id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BindSourceRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Bind a production-safe source provider to a topic.",
        "tags": [
          "source-bindings"
        ]
      }
    },
...
    "/topics/{topicId}/source-bindings/{sourceBindingId}/status": {
      "patch": {
        "operationId": "SourceBindingController_updateStatus",
        "parameters": [
          {
            "in": "path",
            "name": "topicId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "sourceBindingId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "idempotency-key",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Optional Bearer API key. Requires write:source_bindings. If supplied, x-workspace-role is not required.",
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Comma-separated workspace roles. Source binding status updates require owner or admin. Required when Authorization bearer API key is not supplied.",
            "in": "header",
            "name": "x-workspace-role",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-workspace-id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-tenant-id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeSourceBindingStatusRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Pause or resume a source binding.",
        "tags": [
          "source-bindings"
        ]
      }
    },
...
  },
  "servers": [],
...
}